Further shortcut fixes:
Removed previous fix as it just moves the type problems up. Put in the
usual workaround of int(long(x)) to get around casting bug in 64-bit
VC++ compiler.

Shortcut case is preserved and used to decide how to display in the menus.
It is ignored when matching shortcuts. Removed old kludge to control the
capitalization in menus, fixed some of the programs that used lowercase.

Punctuation matches even if shift is held down, so you don't have to
specify shift for punctuation marks (you can if you want to).


git-svn-id: file:///fltk/svn/trunk@6518 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/Widget.h
@@ -22,8 +22,6 @@
 #ifndef fltk_Widget_h
 #define fltk_Widget_h
 
-#include <cstdio>
-
 #include "Style.h"
 #include "Rectangle.h"
 
@@ -103,12 +101,12 @@ public:
   const char *tooltip() const	{ return tooltip_; }
   void	tooltip(const char *t)	{ tooltip_ = t; }
 
-  size_t shortcut() const	;
-  void	shortcut(size_t key)	;
-  bool	add_shortcut(size_t key);
-  bool	remove_shortcut(size_t key);
+  unsigned shortcut() const	;
+  void	shortcut(unsigned key)	;
+  bool	add_shortcut(unsigned key);
+  bool	remove_shortcut(unsigned key);
   void  remove_shortcuts()	;
-  size_t label_shortcut() const;
+  unsigned label_shortcut() const;
   bool	test_label_shortcut() const;
   bool	test_shortcut() const	;
   bool  test_shortcut(bool) const;
